move shouldForceDisableEncryption to /crypto
